ChatBot-API/requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
google-generativeai==0.8.4
langchain==0.3.23
langchain-community==0.3.21
langchain-huggingface==0.1.2
chromadb==0.6.3
sentence-transformers==4.0.2
transformers==4.51.3
torch==2.6.0
pydantic==2.11.3
python-dotenv==1.0.0
